requirements.txt
acme==0.31.0
affine==2.2.2
aiohttp==3.5.4
airtable-local-backup==0.1a1
-e git://github.com/gtalarico/airtable-python-wrapper.git@48b2d806178085b52a31817571e5a1fc3dce4045#egg=airtable_python_wrapper
alabaster==0.7.12
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
Babel==2.6.0
beautifulsoup4==4.7.1
better-exceptions==0.2.2
boto3==1.9.146
botocore==1.12.146
Brlapi==0.6.6
CacheControl==0.12.5
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2019.6.16
cffi==1.11.5
chardet==3.0.4
Click==7.0
click-plugins==1.0.4
cligj==0.5.0
command-not-found==0.3
commonmark==0.9.0
ConfigArgParse==0.11.0
configobj==5.0.0
configparser==3.5.0
coverage==4.5.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
docutils==0.14
enum34==1.1.6
envparse==0.2.0
flake8==3.5.0
Flask==1.1.1
fs==2.4.5
fs-s3fs==1.1.0
funcsigs==1.0.2
future==0.17.1
GDAL==2.4.0
httplib2==0.9.2
hypothesis==3.46.1
idna==2.8
idna-ssl==1.1.0
ijson==2.3
imagesize==1.1.0
ipaddress==1.0.19
iso3166==0.9
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
josepy==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
Loris==2.3.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
mapbox==0.16.1
mapboxcli==0.8.0
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
msgpack==0.6.0
multidict==4.5.2
ndg-httpsclient==0.4.4
netaddr==0.7.19
netifaces==0.10.4
numpy==1.16.4
oauth==1.0.1
olefile==0.45.1
packaging==19.0
pandas==0.24.2
parsedatetime==2.4
pbr==3.1.1
pexpect==4.2.1
Pillow==4.3.0
polyline==1.3.2
protobuf==3.0.0
psycopg2==2.8.3
py==1.5.2
pyasn1==0.4.2
pycairo==1.16.2
pycodestyle==2.3.1
pycparser==2.18
pycrypto==2.6.1
pycups==1.9.73
pyflakes==1.6.0
Pygments==2.4.0
pygobject==3.26.1
PyICU==1.9.8
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.4.0
pyRFC3339==1.0
pytest==3.1.3
pytest-cov==2.5.1
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
python-dotenv==0.10.3
pytz==2019.1
pyxdg==0.25
PyYAML==3.12
rasterio==1.0.18
recommonmark==0.5.0
reportlab==3.4.0
requests==2.22.0
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
responses==0.3.0
ruamel.yaml==0.15.94
s3transfer==0.2.0
SecretStorage==2.3.1
selenium==3.141.0
Shapely==1.5.17
simplejson==3.13.2
six==1.12.0
snowballstemmer==1.2.1
snuggs==1.4.3
soupsieve==1.9.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.5
system-service==0.3
systemd-python==234
tqdm==4.14.0
typing-extensions==3.7.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.25.3
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==0.15.5
xkit==0.0.0
yarl==1.3.0
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
